Avoid most integer conversion warnings from Visual Studio (#5197)
* Avoid unnecessary wchar_t/char conversion in tile_picker_command().

* Internally declare lazymove_delay as a uint8_t since already limited to that range by the save file and when changed by the player.

* Internally declare brand_max, slay_max, and curse_max as uint8_t since the savefile limits them to that range.  Enforce that limit when parsing the data files.

* Tweak wr_ignore() to avoid integer conversion warnings.

* Be explicit about convert the depth from a player or chunk to the origin depth for an object.

* Tweak finish_parse_player_prop() to avoid integer conversion warning from Visual Studio.

* Avoid integer conversion warnings from Visual Studio in expression_add_operations_string().

+++ src/init.h
@@ -37,9 +37,9 @@ struct angband_constants
 	uint16_t s_max;		/**< Maximum number of magic spells */
 	uint16_t pit_max;	/**< Maximum number of monster pit types */
 	uint16_t act_max;	/**< Maximum number of activations for randarts */
-	uint16_t curse_max;	/**< Maximum number of curses */
-	uint16_t slay_max;	/**< Maximum number of slays */
-	uint16_t brand_max;	/**< Maximum number of brands */
+	uint8_t curse_max;	/**< Maximum number of curses */
+	uint8_t slay_max;	/**< Maximum number of slays */
+	uint8_t brand_max;	/**< Maximum number of brands */
 	uint16_t mon_blows_max;	/**< Maximum number of monster blows */
 	uint16_t blow_methods_max;	/**< Maximum number of monster blow methods */
 	uint16_t blow_effects_max;	/**< Maximum number of monster blow effects */
